Alleles with counts in only one position now averaged by 2
When there is a genotype found only once, it must be averaged by 2 pseudogenomes in order to uniform the criteria. The general assumption si that the heterozygous was not detected by one of the pseudogenomes because of the low scoring in the allignment, but the reads were there detected by the other pseudogenome.